LayoutTests/imported/w3c/web-platform-tests/css/css-values/random-in-keyframe-expected.txt


PASS random() is not ignored in keyframe
